+++ packages/navigation/app-navigation/examples/50-theming-example.tsx
@@ -3,11 +3,7 @@ import React, { Fragment } from 'react';
 import { DefaultAppSwitcher } from './shared/AppSwitcher';
 import { DefaultCreate } from './shared/Create';
 import { DefaultHelp } from './shared/Help';
-import { mockEndpoints } from './shared/mock-atlassian-switcher-endpoints';
-import {
-  mockBuiltInNotifications,
-  BuiltInNotifications,
-} from './shared/Notifications';
+import { DefaultNotifications } from './shared/Notifications';
 import { defaultPrimaryItems } from './shared/PrimaryItems';
 import { DefaultProductHome } from './shared/ProductHome';
 import { DefaultProfile } from './shared/Profile';
@@ -15,9 +11,6 @@ import { DefaultSearch } from './shared/Search';
 import { DefaultSettings } from './shared/Settings';
 import { themes } from './shared/themes';
 import { AppNavigation } from '../src';
-
-mockEndpoints('jira');
-mockBuiltInNotifications();
 
 const ThemingExample = () => (
   <div>
@@ -28,7 +21,7 @@ const ThemingExample = () => (
           renderAppSwitcher={DefaultAppSwitcher}
           renderCreate={DefaultCreate}
           renderHelp={DefaultHelp}
-          renderNotifications={BuiltInNotifications}
+          renderNotifications={DefaultNotifications}
           renderProductHome={DefaultProductHome}
           renderProfile={DefaultProfile}
           renderSearch={DefaultSearch}