+++ packages/navigation/app-navigation/src/components/AppSwitcher/index.tsx
@@ -2,21 +2,16 @@ import React from 'react';
 import AppSwitcherIcon from '@atlaskit/icon/glyph/app-switcher';
 
 import { IconButton } from '../IconButton';
-import { TriggerManager } from '../TriggerManager';
 import { AppSwitcherProps } from './types';
 
 export const AppSwitcher = (props: AppSwitcherProps) => {
-  const { tooltip, ...triggerManagerProps } = props;
+  const { tooltip, ...iconButtonProps } = props;
 
   return (
-    <TriggerManager {...triggerManagerProps}>
-      {({ onTriggerClick }) => (
-        <IconButton
-          icon={<AppSwitcherIcon label={tooltip} />}
-          onClick={onTriggerClick}
-          tooltip={tooltip}
-        />
-      )}
-    </TriggerManager>
+    <IconButton
+      icon={<AppSwitcherIcon label={tooltip} />}
+      tooltip={tooltip}
+      {...iconButtonProps}
+    />
   );
 };