+++ packages/navigation/app-navigation/src/components/Notifications/types.ts
@@ -1,26 +1,7 @@
-import { BadgeProps } from '@atlaskit/badge';
-import { TriggerManagerProps } from '../TriggerManager/types';
+import { ComponentType } from 'react';
+import { IconButtonProps } from '../IconButton/types';
 
-type BuiltinBadge = {
-  type: 'builtin';
-  fabricNotificationLogUrl: string;
-  cloudId: string;
-};
-
-type RequiredBadgeProps = Omit<BadgeProps, 'children'>;
-
-type ProvidedBadge = RequiredBadgeProps & {
-  type: 'provided';
-  count: number;
-};
-
-export type NotificationsProps = Omit<
-  TriggerManagerProps,
-  'children' | 'drawerContent'
-> & {
-  badge: BuiltinBadge | ProvidedBadge;
-  drawerContent?: TriggerManagerProps['drawerContent'] | 'builtin';
-  locale?: string;
-  product: string;
+export type NotificationsProps = Omit<IconButtonProps, 'icon' | 'tooltip'> & {
+  badge: ComponentType<{}>;
   tooltip: string;
 };