+++ packages/navigation/app-navigation/src/components/Profile/types.ts
@@ -1,7 +1,3 @@
 import { IconButtonProps } from '../IconButton/types';
-import { TriggerManagerProps } from '../TriggerManager/types';
 
-export type ProfileProps = Omit<TriggerManagerProps, 'children'> & {
-  avatar?: IconButtonProps['icon'];
-  tooltip: string;
-};
+export type ProfileProps = IconButtonProps;