+++ packages/navigation/app-navigation/src/components/Search/index.tsx
@@ -5,7 +5,6 @@ import { Fragment } from 'react';
 
 import { useTheme } from '../../theme';
 import { IconButton } from '../IconButton';
-import { TriggerManager } from '../TriggerManager';
 
 import {
   searchInputContainerCSS,
@@ -29,6 +28,11 @@ const SearchComponent = (props: SearchComponentProps) => {
     onClick && onClick(...args);
   };
 
+  const onInputClick = (...args: any[]) => {
+    // @ts-ignore
+    onClick && onClick(...args);
+  };
+
   return (
     <div css={searchInputContainerCSS}>
       <div css={searchInputIconCSS}>
@@ -38,7 +42,7 @@ const SearchComponent = (props: SearchComponentProps) => {
         css={searchInputCSS(theme)}
         placeholder={text}
         onChange={onChange}
-        onClick={onClick}
+        onClick={onInputClick}
         value=""
       />
     </div>
@@ -46,21 +50,17 @@ const SearchComponent = (props: SearchComponentProps) => {
 };
 
 export const Search = (props: SearchProps) => {
-  const { text, tooltip, ...triggerManagerProps } = props;
+  const { text, tooltip, ...iconButtonProps } = props;
 
   return (
-    <TriggerManager {...triggerManagerProps}>
-      {({ onTriggerClick }) => (
-        <Fragment>
-          <SearchComponent onClick={onTriggerClick} text={text} />
-          <IconButton
-            css={searchIconCSS}
-            icon={<SearchIcon label={tooltip} />}
-            onClick={onTriggerClick}
-            tooltip={tooltip}
-          />
-        </Fragment>
-      )}
-    </TriggerManager>
+    <Fragment>
+      <SearchComponent onClick={iconButtonProps.onClick} text={text} />
+      <IconButton
+        css={searchIconCSS}
+        icon={<SearchIcon label={tooltip} />}
+        tooltip={tooltip}
+        {...iconButtonProps}
+      />
+    </Fragment>
   );
 };